backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-filter>=23.5
django-cors-headers>=4.3.1
drf-spectacular>=0.27.0
psycopg2-binary>=2.9.9
requests>=2.31.0
python-dotenv>=1.0.0 